V0.19.1.0
Added comatose function for when EC or Heat options are ON and Fatal
option is OFF. Now when you run out of EC or overheat with these options
set kerbals will be emergency thawed
and become comatose (tourists) for a period of 5 minutes (can be changed
in the settings menu).
Fixed bug with new pop-up windows when EC is critical or overheat - this
pop-up will now NOT appear if the active vessel is the vessel that has
run out of EC or overheat.